pyproject.toml
# --------------------------------------------------------------
# Project metadata
# --------------------------------------------------------------
[project]
name = "tpr4py"
version = "0.1.0"
description = "Quantitative phase retrieval from bright‑field image stacks"
readme = "README.md"
requires-python = ">=3.9"
license = "GPL-3.0-or-later"
authors = [
    {name = "Moritz Engelhardt", email = "m.l.k.engelahrdt@tudelft.nl"},
]
keywords = ["phase‑imaging", "napari", "microscopy", "image‑processing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
]

# --------------------------------------------------------------
# Runtime dependencies
# --------------------------------------------------------------
dependencies = [
    "numpy>=1.22,<2.0",
    "scikit-image>=0.19.2",
    "napari[all]>=0.4.15",
    "tqdm>=4.64.0",
    "pyyaml>=5.4.0", 
    "numba>=0.56.0 ; platform_system!='Windows'",   # optional JIT on non‑Windows
    "magicgui>=0.7.0",                             # for the optional UI widget                 
]

# --------------------------------------------------------------
# Optional (extra) dependencies
# --------------------------------------------------------------
[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.0.260",          # linting / formatting
    "black>=23.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
napari = [
    "napari>=0.4.15",
    "napari-plugin-engine>=0.2.0",
    "magicgui>=0.7.0",
]

# --------------------------------------------------------------
# Build system (PEP 517)
# --------------------------------------------------------------
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

# --------------------------------------------------------------
# Entry points – CLI & Napari plug‑in
# --------------------------------------------------------------
[project.scripts]               # console‑script entry point
tpr4py = "tpr4py.cli.main:main"


# --------------------------------------------------------------
# Package discovery
# --------------------------------------------------------------
[tool.setuptools.packages.find]
where = ["tpr4py"]
include = ["tpr4py*"]

[tool.setuptools.package-data]
tpr4py = ["napari_plugin/napari.yaml"]

# --------------------------------------------------------------
# Linters / formatters (optional, but handy)
# --------------------------------------------------------------
[tool.ruff]
select = ["E", "F", "W", "C90"]
ignore = ["E501"]               # line length handled by black
target-version = "py39"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true

# --------------------------------------------------------------
# Pytest configuration (optional)
# --------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]